Refactored the whole project
