+++ lib/place/model/place.dart
@@ -1,0 +1,8 @@
+class Place {
+  String name;
+  String where;
+  String type;
+  String steps;
+
+  Place(this.name, this.where, this.type, this.steps);
+}
