+++ lib/place/ui/screens/search_trips.dart
