+++ lib/place/ui/widgets/card_image.dart
