+++ lib/place/ui/widgets/card_image_list.dart
