+++ lib/place/ui/widgets/trips_cupertino.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
-import 'package:trips_app/profile/profile.dart';
-import 'package:trips_app/search_trips.dart';
+import 'package:trips_app/user/ui/screens/profile.dart';
+import 'package:trips_app/place/ui/screens/search_trips.dart';
 
-import 'home_trips.dart';
+import '../screens/home_trips.dart';
 
 class TripsCupertino extends StatelessWidget {
   @override
